KYB_CRAWLERS/KYB-Crawlers-asl_crawlers_prod/requirements.txt
beautifulsoup4==4.11.2
openpyxl==3.1.2
PyPDF2==2.11.1
pandas==1.5.3
pandas_read_xml==0.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
regex==2022.3.15
requests==2.31.0
selenium==4.11.2
shortuuid==1.0.11
webdriver_manager==4.0.1
deep-translator==1.11.0
httpx==0.13.3
asyncio==3.4.3
langdetect==1.0.9
country-converter==0.8.0
googletrans==4.0.0rc1
undetected-chromedriver==3.5.3